refactor: view 내 패키지 분리
