+++ src/main/java/chess/Application.java
@@ -1,8 +1,8 @@
 package chess;
 
 import chess.controller.ChessController;
-import chess.view.InputView;
-import chess.view.OutputView;
+import chess.view.input.InputView;
+import chess.view.output.OutputView;
 
 public class Application {
     public static void main(String[] args) {
