+++ src/main/java/chess/view/input/GameArguments.java
@@ -1,4 +1,4 @@
-package chess.view;
+package chess.view.input;
 
 public record GameArguments(GameCommand gameCommand, MoveArguments moveArguments) {
 }
