+++ src/main/java/chess/view/input/InputView.java
@@ -1,4 +1,4 @@
-package chess.view;
+package chess.view.input;
 
 import java.util.Arrays;
 import java.util.List;
