+++ src/test/java/chess/view/output/PieceShapeTest.java
@@ -1,6 +1,7 @@
-package chess.view;
+package chess.view.output;
 
 import chess.model.piece.*;
+import chess.view.output.PieceShape;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
